Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    double a;
+    float a;
     printf("");
     scanf("%f",&a);
-    printf("You entered: %d",a);
+    printf("You entered: %f",a);
     return 0;
 }
